+++ Models/Headers/Timer.h
@@ -12,7 +12,7 @@ namespace GameNs
 {
     class Timer {
     public:
-        virtual float getDeltaTime()=0;
+        virtual double getDeltaTime()=0;
         virtual void update()=0;
         virtual void reset()=0;
         virtual float getTime()=0;
